LSTM helper fixes; add gradient checks for LSTM layer
Former-commit-id: 3fa50cbd40d0eeedeca7e89e44f98e811de67b3f [formerly a31b4de30e61398bbb1a5fb5efc951bd9b6cf72a]
Former-commit-id: f31d346a791da13b98168b6487b1c6d1eb90d6a8
